Move common DOGM code to headers
