+++ Marlin/src/lcd/HD44780/lcdprint_hd44780.cpp
@@ -19,7 +19,7 @@
 #include "../ultralcd.h"
 #include "../../Marlin.h"
 
-#include "ultralcd_common_HD44780.h"
+#include "ultralcd_HD44780.h"
 
 #include <string.h>
